requirements.txt
Flask==3.0.3
joblib==1.4.2
numpy==2.1.1
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
gunicorn==22.0.0
gevent==24.2.1
gevent-websocket==0.10.1
